+++ src/RcppExports.cpp
@@ -106,6 +106,18 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// pairwise_combination_indices
+NumericMatrix pairwise_combination_indices(unsigned long n, bool self);
+RcppExport SEXP _MESS_pairwise_combination_indices(SEXP nSEXP, SEXP selfSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< unsigned long >::type n(nSEXP);
+    Rcpp::traits::input_parameter< bool >::type self(selfSEXP);
+    rcpp_result_gen = Rcpp::wrap(pairwise_combination_indices(n, self));
+    return rcpp_result_gen;
+END_RCPP
+}
 // qdiag
 NumericVector qdiag(const NumericMatrix& x);
 RcppExport SEXP _MESS_qdiag(SEXP xSEXP) {
@@ -166,6 +178,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_MESS_mfastLmCpp", (DL_FUNC) &_MESS_mfastLmCpp, 3},
     {"_MESS_onemargintest", (DL_FUNC) &_MESS_onemargintest, 2},
     {"_MESS_pairwise_Schur_product", (DL_FUNC) &_MESS_pairwise_Schur_product, 2},
+    {"_MESS_pairwise_combination_indices", (DL_FUNC) &_MESS_pairwise_combination_indices, 2},
     {"_MESS_qdiag", (DL_FUNC) &_MESS_qdiag, 1},
     {"_MESS_quadform", (DL_FUNC) &_MESS_quadform, 4},
     {"_MESS_repmat", (DL_FUNC) &_MESS_repmat, 3},