+++ NEWS.md
@@ -2,6 +2,7 @@
 
 *   Fixed a bug in the documentation for auc for the from argument. Thanks to Maria Meier for finding this.
 *   Added the bin function for fast binning of a numeric vector into equidistant bins
+*   Added the function pairwise_combination_indices to compute all possible pairs of indices.
 
 # MESS 0.5.1